refactor(frontend): audit pinia stores and inline shallow wrappers
Closes #831.

Audited every store under frontend/src/stores/ via the deletion test
(>=2 consumers sharing mutable state -> keep; 1 consumer or pure
pass-through -> delete).

Deleted (single-consumer shallow wrappers, inlined into the page):
- analytics.ts
- apikeys.ts
- chatbot-config.ts
- orgs.ts
- test-sandbox.ts

Kept and documented at the top of each file:
- auth, billing, knowledge-bases, llm-provider-health, llm-providers,
  onboarding, precheck, server-config, voice-sessions, whatsapp,
  workspaces (precheck/server-config/llm-provider-health/passkeys
  already carried explanatory doc-comments).

Deferred (per issue constraint): passkeys.ts — just landed via #777,
left alone as the deletion call is not unambiguous.

Net frontend delta: -832 LOC. All vitest unit specs pass (117),
eslint clean (no new warnings), vue-tsc -b clean.
